Move helpers/ from api.views/ to api/ Implement api.helpers.custom_helper
