+++ app/backend/hiStoryBackend/api/views/auth/views_v1.py
@@ -1,5 +1,5 @@
 from django.http import JsonResponse
-from api.views.helpers import json_response_helper as jrh
+from api.helpers import json_response_helper as jrh
 
 
 def sign_in(request):
