+++ app/backend/hiStoryBackend/api/views/memory_post/views_v1.py
@@ -1,5 +1,5 @@
 from django.http import JsonResponse
-from api.views.helpers import json_response_helper as jrh
+from api.helpers import json_response_helper as jrh
 
 
 def response(request):
